Add files via upload
Drafted SQL for Ressler's Genetics Lab analysis. Query utilizes the rcap_ctau_sample_info_joined_view_new.sql as it's base and can be adapted to the other two labs via filtering the tube_id. There is one outstanding question on whether to expose the dates of last alcohol and drug use or convert these to age (in months). These were not converted in the NDA drug01, but I think these dates should be converted.